backend/requirements.txt
Django==5.0.6
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.32.3
#langchain==0.3.0
#langchain-google-genai==2.1.5
Pillow==10.3.0
google-generativeai==0.8.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.2.2
dj-database-url
djoser
django-cors-headers